web/requirements.txt
# Flask Web Application Dependencies
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0
redis>=5.0.0
rq>=1.15.0
python-dotenv>=1.0.0

# Database (sqlite3 is included in Python standard library)

# Email
sendgrid>=6.11.0

# Date/Time handling for RRULE
python-dateutil>=2.8.0
recurring-ical-events>=2.1.0

# JSON Web Tokens (for admin auth)
PyJWT>=2.8.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-decouple>=3.8

# RRULE scheduling support
icalendar>=5.0.0
croniter>=1.4.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
